Adding logger and logs, refactor structure
