+++ src/main/java/Football/Main.java
@@ -3,10 +3,13 @@ package Football;
 import Controllers.GameController;
 import Entity.Coach;
 import Entity.Player;
+import Enums.PlayerRole;
 import Exceptions.NotNegativeDateException;
 import Exceptions.NotNegativeScoreException;
+import Interfaces.IAction;
 
 public class Main {
+
     public static void main(String[] args) {
         GameController gameController = new GameController();
         Score score = new Score(2, 1);
