+++ src/main/java/Interfaces/IAction.java
@@ -1,4 +1,4 @@
-package Football;
+package Interfaces;
 
 public interface IAction {
     String action();
